requirements.txt
Django==1.10.4
google-api-python-client==1.5.5
httplib2==0.9.2
Markdown==2.6.7
oauth2client==4.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyYAML==3.12
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
